feat: docker build module

+++ deploy/update.sh
@@ -1,3 +1,6 @@
-docker-compose stop
+docker-compose down
+if $2
+then
 docker rmi $(docker images app:$2) -f
+fi
 docker-compose up -d
